Minor Site Server tiny up
