+++ src/site/server/config.ts
@@ -7,8 +7,12 @@ import state from '../../common/state'
 export let config: Config = null
 
 export const loadConfig = async () => {
-  logStep(`Site Server is loading config from ${process.env[CONFIG_FILE_PATH_ENV_VAR_NAME]}`, true)
-  config = await getConfig(process.env[CONFIG_FILE_PATH_ENV_VAR_NAME])
+  const configFilePath = process.env[CONFIG_FILE_PATH_ENV_VAR_NAME]
+  if (configFilePath != null)
+    logStep(`Site Server is loading config from ${configFilePath}`, true)
+  else
+    logStep('Site Server is loading default config as no config file path has been provided.', true)
+  config = await getConfig(configFilePath)
   // If verbose mode has not already been enabled, refer to the value from config
   if (!state.verbose)
     state.verbose = config.verbose
