+++ src/site/server/types.ts
@@ -1,0 +1,3 @@
+import express from 'express'
+
+export type ExpressApp = ReturnType<typeof express>
